refactor: extract youtube into a standalone file
